refactored and renamed files
